Move it all over to DatabaseTable
