+++ modules/datastore/src/Plugin/QueueWorker/Import.php
@@ -110,7 +110,6 @@ class Import extends QueueWorkerBase implements ContainerFactoryPluginInterface 
     $timeout = (int) $plugin_definition['cron']['lease_time'];
     $this->setConnectionTimeout($datastoreConnection, $timeout);
     $this->setConnectionTimeout($defaultConnection, $timeout);
-    $this->setInnodbMode($datastoreConnection);
   }
 
   /**
@@ -141,16 +140,6 @@ class Import extends QueueWorkerBase implements ContainerFactoryPluginInterface 
   protected function setConnectionTimeout(Connection $connection, int $timeout): void {
     $command = 'SET SESSION wait_timeout = ' . $timeout;
     $connection->query($command);
-  }
-
-  /**
-   * Disable InnoDB strict mode for the given database connection.
-   *
-   * @param \Drupal\Core\Database\Connection $connection
-   *   Database connection instance.
-   */
-  public function setInnodbMode(Connection $connection): void {
-    $connection->query("SET SESSION innodb_strict_mode=OFF");
   }
 
   /**
