+++ modules/datastore/tests/src/Functional/Storage/DatabaseTableTest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Drupal\Tests\datastore\Functional\Service;
+namespace Drupal\Tests\datastore\Functional\Storage;
 
+use Dkan\Datastore\Resource;
+use Drupal\datastore\Storage\DatabaseTable;
 use Drupal\Tests\common\Traits\CleanUp;
 use Drupal\Tests\common\Traits\GetDataTrait;
-
 use weitzman\DrupalTestTraits\ExistingSiteBase;
 
 /**
@@ -13,16 +14,9 @@ use weitzman\DrupalTestTraits\ExistingSiteBase;
  * @package Drupal\Tests\datastore\Functional
  * @group datastore
  */
-class ImportTest extends ExistingSiteBase {
+class DatabaseTableTest extends ExistingSiteBase {
   use GetDataTrait;
   use CleanUp;
-
-  public function setUp(): void {
-    parent::setUp();
-
-    // Initialize services.
-    $this->queueWorkerManager = \Drupal::service('plugin.manager.queue_worker');
-  }
 
   /**
    * Test that setInnodbMode() turns strict mode off for datastore but not
@@ -33,14 +27,22 @@ class ImportTest extends ExistingSiteBase {
     $result = $connection->query("SHOW SESSION VARIABLES LIKE 'innodb_strict_mode'")->fetchObject();
     $this->assertEquals($result->Value, "ON");
 
-    $importWorker = $this->queueWorkerManager->createInstance("datastore_import");
-    $importWorker->setInnodbMode($connection);
+    $resource = new Resource('123', '/tmp', 'text/csv');
+
+    $databaseTable = new DatabaseTable($connection, $resource);
+    $databaseTable->setInnodbMode("OFF");
 
     $result = $connection->query("SHOW SESSION VARIABLES LIKE 'innodb_strict_mode'")->fetchObject();
     $this->assertEquals($result->Value, "OFF");
 
+    // Other database connection not affected.
     $connection2 = \Drupal::service('database');
     $result = $connection2->query("SHOW SESSION VARIABLES LIKE 'innodb_strict_mode'")->fetchObject();
+    $this->assertEquals($result->Value, "ON");
+
+    // Safe mode can be turned back on.
+    $databaseTable->setInnodbMode("ON");
+    $result = $connection->query("SHOW SESSION VARIABLES LIKE 'innodb_strict_mode'")->fetchObject();
     $this->assertEquals($result->Value, "ON");
   }
